Create new controller and new service for task navigation
