+++ backend/src/Designer/Services/Interfaces/IAppDevelopmentService.cs
@@ -115,13 +115,6 @@ namespace Altinn.Studio.Designer.Services.Interfaces
         public Task<LayoutSetsModel> GetLayoutSetsExtended(AltinnRepoEditingContext altinnRepoEditingContext, CancellationToken cancellationToken = default);
 
         /// <summary>
-        /// Get task navigation
-        /// </summary>
-        /// <param name="altinnRepoEditingContext">An <see cref="AltinnRepoEditingContext"/>.</param>
-        /// <param name="cancellationToken">A <see cref="CancellationToken"/> that observes if operation is canceled.</param>
-        public Task<List<TaskNavigationModel>> GetTaskNavigation(AltinnRepoEditingContext altinnRepoEditingContext, CancellationToken cancellationToken = default);
-
-        /// <summary>
         /// Gets a layoutSet config.
         /// </summary>
         /// <param name="altinnRepoEditingContext">An <see cref="AltinnRepoEditingContext"/>.</param>
